requirements.txt
```
pillow
````